+++ app/src/main/java/com/example/revivo/data/local/model/ExerciseLog.java
@@ -1,18 +1,17 @@
 package com.example.revivo.data.local.model;
 
-
 public class ExerciseLog {
     private long id;
     private long userId;
     private String exerciseId;
     private String exerciseName;
-    private int durationMin;
+    private double durationMin;
     private String completedAt;
 
     public ExerciseLog() {}
 
     public ExerciseLog(long id, long userId, String exerciseId,
-                       String exerciseName, int durationMin, String completedAt) {
+                       String exerciseName, double durationMin, String completedAt) {
         this.id = id;
         this.userId = userId;
         this.exerciseId = exerciseId;
@@ -33,10 +32,9 @@ public class ExerciseLog {
     public String getExerciseName() { return exerciseName; }
     public void setExerciseName(String exerciseName) { this.exerciseName = exerciseName; }
 
-    public int getDurationMin() { return durationMin; }
-    public void setDurationMin(int durationMin) { this.durationMin = durationMin; }
+    public double getDurationMin() { return durationMin; }
+    public void setDurationMin(double durationMin) { this.durationMin = durationMin; }
 
     public String getCompletedAt() { return completedAt; }
     public void setCompletedAt(String completedAt) { this.completedAt = completedAt; }
 }
-
